automate payment and returning emails;(#27) (#28)
Adds the payment and returning emails to curateTodaysEmail(). Payment emails will send through the first 3 games to avoid extra spam if they paid and their status was not marked. Returning emails send as long as there are atleast 3 games on the schedule and 3 games or fewer left for the season.